* SysLinker
- CommonFunc

1. ROM 파일 변환 후 프로젝트 번호 갱신을 위해 중계기 일람표를 찾아야 하는데 이를 위해 특정 폴더 내의 특정 파일을 찾는 공용 매서드 추가 (단, 폴더 내 폴더 검색은 안됨)

- Defines

1. CCTV 정보가 삭제됨에 따라 이에 따른 정의 삭제

2. F4 / F3 수신기 LCD 최대 문자열 정의 추가

3. ROM 파일 변환 시 관리자 모드 암호 정의 추가

4. F4 추가 입력 타입 인덱스 수정 (힐스테이트남산 현장 확인)

5. CCTV 정보 삭제 후 Ficon에서 처리를 위해 [입력 타입], [설비명]에 [CCTV]항목 정의 추가

6. 프로젝트 시트 엑셀 ROW 순서 정의 변경

7. F4APPENDIX_INFO에 PROJECT_INFO 포함

- NewDBManager

1. F4 입력타입의 인덱스를 DB에 고정으로 입력하고 만약 기존에 해당 인덱스를 사용하는 입력타입이 있다고 하더라도 Update하도록 함 (입력 타입은 현장과 프로그램팀의 약속이 된 규칙이므로)

2. CCTV 의 [설비명] 인덱스는 기존 [설비명] 마지막 인덱스 다음 인덱스로 입력되도록 함

3. CCTV 관련 매서드 삭제

- NewExcelManager

1. F4 추가 시트 존재 여부 플래그 추가

2. CCTV 관련 매서드 삭제

3. ROM 파일 변환 후 중계기 일람표 상 프로젝트 자동 갱신 매서드 추가

- NewInfo

1. PROJECT_INFO가 F4APPENDIX_INFO에 포함됨에 따라 PROJECT_INFO 인스턴스 삭제

- res

1. 리본 메뉴 변경에 따른 이미지 파일들 추가 및 설정

- DlgAdminMode

1. ROM 파일 변환 시 관리자 모드 암호 입력 받는 창 추가

- DlgLogin

1. 항상 1.0으로 잘못된 프로젝트 버전 번호를 표현하는 부분을 안보이게 처리

- FormAutoMake

1. 주석 삭제

- FormLoadRelayTable

1. 주석 삭제

2. F4 추가 테이블의 경우 중계기 일람표 상 시트에서 F4추가 정보 시트가 존재할 경우에만 생성

- MainFrm

1. 테스트 코드 삭제

2. ROM 파일 변환 전 관리자 모드로 실행할 지 여부를 묻도록 추가

3. 연동데이터 생성 후 수신기 타입 중 하나라도 F4 타입이 있으면 프로젝트 버전 번호 관련 처리를 진행하고 그렇지 않으면 의미가 없으므로 하지 않음

- PropPageDevice

1. 프로젝트 생성 중 중계기 일람표 선택 시 중계기 일람표 선택 -> 리스트 삭제 -> 선택을 진행했을 때 이미 선택한 번호라고 나오면 선택이 되지 않는 오류 수정

- PropPageProjectInfo

1. 프로젝트 이름을 중계기 일람표 상의 정보를 사용하기로 했으므로 SLP4 프로젝트 이름을 적용하는 코드 삭제 (궁극적으로는 일치시키는 방법으로 추후에 나아가야 함)

- RelayTableData

1. 중계기 일람표 파싱 시 F4 추가 시트 존재 여부 판단 플래그로 F4 추가 정보가 있는지를 확인하고 여러 개의 중계기 일람표가 있을 경우 F4 추가 정보 시트는 한번만 처리하도록 함

2. 기존 프로젝트 로드가 완료가 될 때 마지막으로 F4 추가 정보 테이블이 존재하면 F4 추가 정보를 로드하도록 함

3. ROM 파일명 생성 시 F4 타입은 프로젝트 번호 정보가 들어간 ROM 파일명으로 생성

4. LCD 문자열 처리 시 F4 타입은 80바이트, F3 타입은 40바이트까지 처리하도록 함

5. ROM 파일 생성 시 입력타입 개수가 추가됨에 따라 최대 57개까지 처리하도록 추가

- SysLinker

1. 테스트 코드 삭제

2. 프로젝트 이름을 중계기 일람표 상의 정보를 사용하기로 했으므로 SLP4 프로젝트 이름을 적용하는 코드 삭제 (궁극적으로는 일치시키는 방법으로 추후에 나아가야 함)

3. [ROM 파일 역변환] 버튼 비활성화

4. 새 프로젝트 DB 생성이 완료될 때 앞서 중계기 일람표 파싱 시 F4 추가 정보가 존재하면 F4 추가 정보 테이블을 생성하도록 함

5. 새 프로젝트 생성 시 DB Open이 끝나면 DB관련 처리를 하기 이전에 F4 추가 입력 타입을 입력해서 이후 루틴에 같이 처리되도록 함

6. 프로젝트가 닫히면 F4 관련 메모리를 초기화

+++ SLP3.5Func/NewDBManager.cpp
@@ -22,7 +22,7 @@ BOOL CNewDBManager::CheckAndInsertEquipmentNewInputType()
 	if (m_pDB == nullptr)
 		return FALSE;
 
-	for (int i = AN광전교차A; i <= 광센서감지기; i++)
+	for (int i = NEW_EQUIPMENT_INPUT_TYPE::AN정온교차A; i <= NEW_EQUIPMENT_INPUT_TYPE::CCTV; i++)
 	{
 		CString strEquip = _T("");
 		strEquip.Format(_T("%s"), g_lpszNewEquipmentInputType[i]);
@@ -41,11 +41,94 @@ BOOL CNewDBManager::CheckAndInsertEquipmentNewInputType()
 		if (nRow == 0)
 		{
 			strQuery.Format(_T("INSERT INTO TB_EQUIP_MST (EQ_ID, EQ_TYPE, EQ_NAME, EQ_DESC, EQ_SYMBOL, ADD_USER) VALUES (%d, 1, \'%s\', \'%s\', \'basic.bmp\', \'admin\')"), i, strEquip, strEquip);
-			if (!m_pDB->ExecuteSql(strQuery))
+			if (m_pDB->ExecuteSql(strQuery))
+			{
+				Log::Trace("TB_EQUIP_MST INSERT QUERY Succeeded! Input Type : %s", CCommonFunc::WCharToChar(strEquip.GetBuffer(0)));
+			}
+			else
 			{
 				Log::Trace("TB_EQUIP_MST INSERT QUERY Failed! - Input Type : %s", CCommonFunc::WCharToChar(strEquip.GetBuffer(0)));
 				return FALSE;
 			}
+		}
+		else if (nRow == 1)	// 만약 존재하면 업데이트하고 로그 남김
+		{
+			strQuery.Format(_T("UPDATE TB_EQUIP_MST SET EQ_NAME = \'%s\', EQ_DESC = \'%s\' WHERE EQ_TYPE = 1 AND EQ_ID = %d"), strEquip, strEquip, i);
+			if (m_pDB->ExecuteSql(strQuery))
+			{
+				Log::Trace("TB_EQUIP_MST UPDATE QUERY Succeeded! - Input Type : %s", CCommonFunc::WCharToChar(strEquip.GetBuffer(0)));
+			}
+			else
+			{
+				Log::Trace("TB_EQUIP_MST UPDATE QUERY Failed! - Input Type : %s", CCommonFunc::WCharToChar(strEquip.GetBuffer(0)));
+				return FALSE;
+			}
+		}
+		else
+		{
+			Log::Trace("There is more than one row ! - Input Type : %s Please Check TB_EQUIP_MST table out", CCommonFunc::WCharToChar(strEquip.GetBuffer(0)));
+			return FALSE;
+		}
+	}
+
+	//입력 설비명 CCTV 추가 - 기본 DB (gfs_base)에는 52번으로 넣지만 기존 DB로드 시 CCTV 이름을 추가할 경우 기존에 없으면 맨 마지막 번으로 넣음
+	for (int j = NEW_EQUIPMENT_INPUT_NAME::CCTV; j <= NEW_EQUIPMENT_INPUT_NAME::CCTV; j++)
+	{
+		CString strName = _T("");
+		strName.Format(_T("%s"), g_lpszNewEquipmentInputName[j]);
+
+		CString strQuery = _T("");
+		strQuery.Format(_T("SELECT * FROM TB_EQUIP_MST WHERE EQ_TYPE=2 AND EQ_NAME=\'%s\' AND EQ_DESC=\'%s\'"), strName, strName);
+
+		if (!m_pDB->OpenQuery(strQuery))
+		{
+			Log::Trace("TB_EQUIP_MST SELECT QUERY Failed! - Input Name : %s", CCommonFunc::WCharToChar(strName.GetBuffer(0)));
+			return FALSE;
+		}
+
+		int nRow = -1;
+		nRow = m_pDB->GetRecordCount();
+		if (nRow == 0)
+		{
+			//기존에 입력명이 몇 개 있는지 확인
+			strQuery.Format(_T("SELECT * FROM TB_EQUIP_MST WHERE EQ_TYPE=2"));
+
+			if (!m_pDB->OpenQuery(strQuery))
+			{
+				Log::Trace("TB_EQUIP_MST SELECT QUERY (EQ_TYPE=2 COUNT) Failed! - Input Name : %s", CCommonFunc::WCharToChar(strName.GetBuffer(0)));
+				return FALSE;
+			}
+
+			nRow = m_pDB->GetRecordCount();
+
+			strQuery.Format(_T("INSERT INTO TB_EQUIP_MST (EQ_ID, EQ_TYPE, EQ_NAME, EQ_DESC, EQ_SYMBOL, ADD_USER) VALUES (%d, 2, \'%s\', \'%s\', \'basic.bmp\', \'admin\')"), nRow + 1, strName, strName);
+			if (!m_pDB->ExecuteSql(strQuery))
+			{
+				Log::Trace("TB_EQUIP_MST INSERT QUERY Failed! - Input Name : %s", CCommonFunc::WCharToChar(strName.GetBuffer(0)));
+				return FALSE;
+			}
+		}
+		else if (nRow == 1)	// 만약 존재하면 아무 것도 안함
+		{
+// 			int nEQID = -1;
+// 			m_pDB->GetFieldValue(_T("EQ_ID"), nEQID);
+// 			ASSERT(nEQID > 0);
+// 
+// 			strQuery.Format(_T("UPDATE TB_EQUIP_MST SET EQ_NAME = \'%s\', EQ_DESC = \'%s\' WHERE EQ_TYPE = 2 AND EQ_ID = %d"), strName, strName, nEQID);
+// 			if (m_pDB->ExecuteSql(strQuery))
+// 			{
+// 				Log::Trace("TB_EQUIP_MST UPDATE QUERY Succeeded! - Input Name : %s", CCommonFunc::WCharToChar(strName.GetBuffer(0)));
+// 			}
+// 			else
+// 			{
+// 				Log::Trace("TB_EQUIP_MST UPDATE QUERY Failed! - Input Name : %s", CCommonFunc::WCharToChar(strName.GetBuffer(0)));
+// 				return FALSE;
+// 			}
+		}
+		else
+		{
+			Log::Trace("There is more than one row ! - Input Name : %s Please check TB_EQUIP_MST table out", CCommonFunc::WCharToChar(strName.GetBuffer(0)));
+			return FALSE;
 		}
 	}
 
@@ -77,11 +160,6 @@ BOOL CNewDBManager::CheckAndCreateF4DBTables()
 			case TB_UNIT_TYPE:
 			{
 				strQuery = GetQueryStringCreateUnitTypeTable();
-				break;
-			}
-			case TB_CCTV_INFO:
-			{
-				strQuery = GetQueryStringCreateCCTVInfoTable();
 				break;
 			}
 			case TB_PROJECT_INFO:
@@ -157,25 +235,6 @@ CString CNewDBManager::GetQueryStringCreateUnitTypeTable()
 	return strQuery;
 }
 
-CString CNewDBManager::GetQueryStringCreateCCTVInfoTable()
-{
-	CString strQuery = _T("");
-
-	strQuery = _T("CREATE TABLE [dbo].[TB_CCTV_INFO]( ");
-	strQuery += _T("[NO] int NOT NULL PRIMARY KEY, ");
-	strQuery += _T("[CCTV_TYPE] tinyint NOT NULL, ");
-	strQuery += _T("[COMPANY_TYPE] tinyint NOT NULL, ");
-	strQuery += _T("[IP] varchar(16), ");
-	strQuery += _T("[PORT] int, ");
-	strQuery += _T("[URL] varchar(2084), ");
-	strQuery += _T("[CAMERA_COUNT] int, ");
-	strQuery += _T("[ID] varchar(20), ");
-	strQuery += _T("[PASSWORD] varchar(20), ");
-	strQuery += _T("[RESERVED] bigint )");
-
-	return strQuery;
-}
-
 CString CNewDBManager::GetQueryStringCreateProjectInfoTable()
 {
 	CString strQuery = _T("");
@@ -196,9 +255,6 @@ BOOL CNewDBManager::InsertDatasIntoF4DBTables()
 		return FALSE;
 
 	if (!InsertDataIntoUnitTypeTable())
-		return FALSE;
-
-	if (!InsertDataIntoCCTVInfoTable())
 		return FALSE;
 
 	if (!InsertDataIntoProjectInfoTable())
@@ -285,88 +341,34 @@ BOOL CNewDBManager::InsertDataIntoUnitTypeTable()
 	return TRUE;
 }
 
-BOOL CNewDBManager::InsertDataIntoCCTVInfoTable()
-{
-	CString strQuery = _T("DELETE FROM [dbo].[TB_CCTV_INFO]");
-	if (!m_pDB->ExecuteSql(strQuery))
-	{
-		Log::Trace("[%s] query failed!", CCommonFunc::WCharToChar(strQuery.GetBuffer(0)));
-	}
-
-	for (int i = 0; i < MAX_CCTV_COUNT; i++)
-	{
-		int nNum = -1;
-		int nCCTVType = 0;
-		int nCompanyType = 0;
-		CString strIP = _T("");
-		int nPort = -1;
-		CString strURL = _T("");
-		int nCameraCount = -1;
-		CString strID = _T("");
-		CString strPassword = _T("");
-		double nReserved = 0;
-
-		//번호는 1베이스
-		nNum = i + 1;
-
-		nCCTVType = CNewInfo::Instance()->m_fi.cctvInfo[i].cctvType;
-		if(nCCTVType == 0)
-		{
-			Log::Trace("The total number of inserted CCTV information is %d", i);
-			break;
-		}
-
-		nCompanyType = CNewInfo::Instance()->m_fi.cctvInfo[i].companyType;
-
-		strIP = CCommonFunc::CharToWCHAR(CNewInfo::Instance()->m_fi.cctvInfo[i].ip);
-
-		nPort = CNewInfo::Instance()->m_fi.cctvInfo[i].port;
-
-		strURL = CCommonFunc::CharToWCHAR(CNewInfo::Instance()->m_fi.cctvInfo[i].url);
-
-		nCameraCount = CNewInfo::Instance()->m_fi.cctvInfo[i].cameraCount;
-
-		strID = CCommonFunc::CharToWCHAR(CNewInfo::Instance()->m_fi.cctvInfo[i].id);
-
-		strPassword = CCommonFunc::CharToWCHAR(CNewInfo::Instance()->m_fi.cctvInfo[i].password);
-
-		nReserved = CNewInfo::Instance()->m_fi.cctvInfo[i].reserved;
-
-		strQuery.Format(_T("INSERT INTO TB_CCTV_INFO (NO, CCTV_TYPE, COMPANY_TYPE, IP, PORT, URL, CAMERA_COUNT, ID, PASSWORD, RESERVED) VALUES (%d, %d, %d, \'%s\', %d, \'%s\', %d, \'%s\', \'%s\', %d)"), 
-			nNum, nCCTVType, nCompanyType, strIP, nPort, strURL, nCameraCount, strID, strPassword, nReserved);
-
-		if (!m_pDB->ExecuteSql(strQuery))
-		{
-			Log::Trace("[%s] query failed!", CCommonFunc::WCharToChar(strQuery.GetBuffer(0)));
-			return FALSE;
-		}
-	}
-
-	return TRUE;
-}
-
 BOOL CNewDBManager::InsertDataIntoProjectInfoTable()
 {
+	//만약 모든 수신기가 F3라서 프로젝트명이 존재하지 않으면 DB Insert를 진행하지 않음
+	CString strProjectName = _T("");
+	strProjectName = CCommonFunc::CharToWCHAR(CNewInfo::Instance()->m_fi.projectInfo.projectName);
+	if(strProjectName.IsEmpty())
+	{ 
+		Log::Trace("There in no project information!");
+		return TRUE;
+	}
+
 	CString strQuery = _T("DELETE FROM [dbo].[TB_PROJECT_INFO]");
 	if (!m_pDB->ExecuteSql(strQuery))
 	{
 		Log::Trace("[%s] query failed!", CCommonFunc::WCharToChar(strQuery.GetBuffer(0)));
 	}
 
-	CString strProjectName = _T("");
 	int nModuleTableVerNum = -1;
 	int nLinkedDataVerNum = -1;
 	int bAuthorized = -1;
 
-	strProjectName = CCommonFunc::CharToWCHAR(CNewInfo::Instance()->m_pi.projectName);
-
-	nModuleTableVerNum = CNewInfo::Instance()->m_pi.moduleTableVerNum;
+	nModuleTableVerNum = CNewInfo::Instance()->m_fi.projectInfo.moduleTableVerNum;
 	ASSERT(nModuleTableVerNum != -1);
 
-	nLinkedDataVerNum = CNewInfo::Instance()->m_pi.linkedDataVerNum;
+	nLinkedDataVerNum = CNewInfo::Instance()->m_fi.projectInfo.linkedDataVerNum;
 	ASSERT(nLinkedDataVerNum != -1);
 
-	bAuthorized = CNewInfo::Instance()->m_pi.authorized;
+	bAuthorized = CNewInfo::Instance()->m_fi.projectInfo.authorized;
 	ASSERT(bAuthorized != -1);
 
 	strQuery.Format(_T("INSERT INTO TB_PROJECT_INFO (PROJECT_NAME, MODULE_TABLE_VER_NUM, LINKED_DATA_VER_NUM, AUTHORIZED) VALUES (\'%s\', %d, %d, %d)"),
@@ -377,6 +379,8 @@ BOOL CNewDBManager::InsertDataIntoProjectInfoTable()
 		Log::Trace("[%s] query failed!", CCommonFunc::WCharToChar(strQuery.GetBuffer(0)));
 		return FALSE;
 	}
+
+	return TRUE;
 }
 
 BOOL CNewDBManager::GetDataFromF4DBTables()
@@ -385,9 +389,6 @@ BOOL CNewDBManager::GetDataFromF4DBTables()
 		return FALSE;
 
 	if (!GetDataFromUnitTypeTable())
-		return FALSE;
-
-	if (!GetDataFromCCTVInfoTable())
 		return FALSE;
 
 	if (!GetDataFromProjectInfoTable())
@@ -428,8 +429,8 @@ BOOL CNewDBManager::GetDataFromFACPTypeTable()
 	}
 	else
 	{
-		Log::Trace("The number of records in the TB_FACP_TYPE table is 0.");
-		return FALSE;
+// 		Log::Trace("The number of records in the TB_FACP_TYPE table is 0.");
+// 		return FALSE;
 	}
 
 	return TRUE;
@@ -470,67 +471,8 @@ BOOL CNewDBManager::GetDataFromUnitTypeTable()
 	}
 	else
 	{
-		Log::Trace("The number of records in the TB_UNIT_TYPE table is 0.");
-		return FALSE;
-	}
-
-	return TRUE;
-}
-
-BOOL CNewDBManager::GetDataFromCCTVInfoTable()
-{
-	CString strQuery = _T("SELECT * FROM [dbo].[TB_CCTV_INFO]");
-
-	if (!m_pDB->OpenQuery(strQuery))
-	{
-		Log::Trace("[%s] query failed!", CCommonFunc::WCharToChar(strQuery.GetBuffer(0)));
-		return FALSE;
-	}
-
-	int nRecordCount = -1;
-	nRecordCount = m_pDB->GetRecordCount();
-	if (nRecordCount > 0)	// CCTV는 없을 수도 있음
-	{
-		for (int i = 0; i < nRecordCount; i++)
-		{
-			int nNum = -1;
-			int nCCTVType = 0;
-			int nCompanyType = 0;
-			CString strIP = _T("");
-			int nPort = 0;
-			CString strURL = _T("");
-			int nCameraCount = 0;
-			CString strID = _T("");
-			CString strPassword = _T("");
-			double nReserved = 0;
-
-			m_pDB->GetFieldValue(_T("NO"), nNum);
-			ASSERT(nNum > -1);
-			m_pDB->GetFieldValue(_T("CCTV_TYPE"), nCCTVType);
-			m_pDB->GetFieldValue(_T("COMPANY_TYPE"), nCompanyType);
-			m_pDB->GetFieldValue(_T("IP"), strIP);
-			m_pDB->GetFieldValue(_T("PORT"), nPort);
-			m_pDB->GetFieldValue(_T("URL"), strURL);
-			m_pDB->GetFieldValue(_T("CAMERA_COUNT"), nCameraCount);
-			m_pDB->GetFieldValue(_T("ID"), strID);
-			m_pDB->GetFieldValue(_T("PASSWORD"), strPassword);
-			m_pDB->GetFieldValue(_T("RESERVED"), nReserved);
-
-
-			ASSERT(nNum > 0);
-			nNum -= 1;	// 번호는 1베이스, 인덱스는 0베이스
-			CNewInfo::Instance()->m_fi.cctvInfo[nNum].cctvType = nCCTVType;
-			CNewInfo::Instance()->m_fi.cctvInfo[nNum].companyType = nCompanyType;
-			strcpy_s(CNewInfo::Instance()->m_fi.cctvInfo[nNum].ip, CCommonFunc::WCharToChar(strIP.GetBuffer(0)));
-			CNewInfo::Instance()->m_fi.cctvInfo[nNum].port = nPort;
-			strcpy_s(CNewInfo::Instance()->m_fi.cctvInfo[nNum].url, CCommonFunc::WCharToChar(strURL.GetBuffer(0)));
-			CNewInfo::Instance()->m_fi.cctvInfo[nNum].cameraCount = nCameraCount;
-			strcpy_s(CNewInfo::Instance()->m_fi.cctvInfo[nNum].id, CCommonFunc::WCharToChar(strID.GetBuffer(0)));
-			strcpy_s(CNewInfo::Instance()->m_fi.cctvInfo[nNum].password, CCommonFunc::WCharToChar(strPassword.GetBuffer(0)));
-			CNewInfo::Instance()->m_fi.cctvInfo[nNum].reserved = nReserved;
-
-			m_pDB->MoveNext();
-		}
+// 		Log::Trace("The number of records in the TB_UNIT_TYPE table is 0.");
+// 		return FALSE;
 	}
 
 	return TRUE;
@@ -565,15 +507,15 @@ BOOL CNewDBManager::GetDataFromProjectInfoTable()
 		ASSERT(nLinkedDataVerNum > -1);
 		ASSERT(nAuthorized > -1);
 
-		strcpy_s(CNewInfo::Instance()->m_pi.projectName, CCommonFunc::WCharToChar(strProjectName.GetBuffer(0)));
-		CNewInfo::Instance()->m_pi.moduleTableVerNum = nModuleTableVerNum;
-		CNewInfo::Instance()->m_pi.linkedDataVerNum = nLinkedDataVerNum;
-		CNewInfo::Instance()->m_pi.authorized = nAuthorized;
+		strcpy_s(CNewInfo::Instance()->m_fi.projectInfo.projectName, CCommonFunc::WCharToChar(strProjectName.GetBuffer(0)));
+		CNewInfo::Instance()->m_fi.projectInfo.moduleTableVerNum = nModuleTableVerNum;
+		CNewInfo::Instance()->m_fi.projectInfo.linkedDataVerNum = nLinkedDataVerNum;
+		CNewInfo::Instance()->m_fi.projectInfo.authorized = nAuthorized;
 	}
 	else
 	{
-		Log::Trace("The number of records in the TB_PROJECT_INFO table is 0.");
-		return FALSE;
+// 		Log::Trace("The number of records in the TB_PROJECT_INFO table is 0.");
+// 		return FALSE;
 	}
 
 	return TRUE;
